Finished the Assignments Grading spec. Creation of the assignment duplication test case spec.
Added the set_options method to the PageHelper module. Updated the data object classes to use this method.

Added requires method to pagehelper, updated data object classes to use this method.

Adding more stuff for the assignment grading spec.  Renamed a page object class (AssignmentView) but didn't update all the specs, so this probably broke a bunch of stuff.

Starting to consolidate the AssignmentView and AssignmentStudent classes.  They overlap so much that it'll just be simpler to combine them.

Finished the Assignments Grading spec.

Finished the Assignments Grading spec.